primeros avances de permisos de usuarios en todas las aplicaciones. correcciones en templaes varios

+++ apps/generales/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from django.views.generic import CreateView, UpdateView, DeleteView
+from django.views.generic import CreateView, UpdateView, DeleteView,DetailView
 from django.http import HttpResponseRedirect, JsonResponse
 from django.shortcuts import redirect
 from django_tables2.views import SingleTableMixin
@@ -7,10 +7,14 @@ from django_filters.views import FilterView
 
 from django_tables2.views import SingleTableView
 from django_tables2.export.views import ExportMixin
-# Create your views here.
-class GenericAltaView(CreateView):
-	botones = dict()
+from django.contrib.auth.mixins import LoginRequiredMixin,PermissionRequiredMixin
+from django.contrib import messages
 
+
+class GenericAltaView(LoginRequiredMixin,PermissionRequiredMixin,CreateView):
+#class GenericAltaView(CreateView):
+	botones = dict()
+	
 	def form_valid(self,form):
 		self.object = form.save()
 		if self.object != None:
@@ -29,6 +33,10 @@ class GenericAltaView(CreateView):
 		return_path = request.POST.get('guardar', self.success_url)
 		return redirect(return_path)
 
+	def handle_no_permission(self):
+		messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+		return redirect(self.redirect_url)
+
 	def get_context_data(self, **kwargs):
 		context = super(GenericAltaView, self).get_context_data(**kwargs)
 		context['return_path'] = self.request.GET.get('return_path', self.success_url)
@@ -36,30 +44,46 @@ class GenericAltaView(CreateView):
 		context['botones'] = self.botones
 		return context
 
-class GenericModificacionView(UpdateView):
+class GenericDetalleView(LoginRequiredMixin,PermissionRequiredMixin,DetailView):
+#class GenericDetalleView(DetailView):
+	def handle_no_permission(self):
+		messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+		return redirect(self.redirect_url)
 
+class GenericModificacionView(LoginRequiredMixin,PermissionRequiredMixin,UpdateView):
 
-	def get_context_data(self, **kwargs):
-		context = super(GenericModificacionView, self).get_context_data(**kwargs)
-		context['nombreForm'] = self.nombre_form
-		context['botones'] = self.botones
-		return context
+	def handle_no_permission(self):
+		messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+		return redirect(self.redirect_url)
 
-class GenericEliminarView(DeleteView):
+class GenericEliminarView(LoginRequiredMixin,DeleteView):
 	def delete(self, request, *args, **kwargs):
-		try:
-			self.object = self.get_object()
-			self.object.delete()
+		if request.user.has_perm(self.permission_required):
+			try:
+				self.object = self.get_object()
+				self.object.delete()
 
+				return JsonResponse({
+					"success": True
+				})
+			except Error as e:
+				return JsonResponse({
+					"success": False,
+					"message": ('error','Este documento no se puede eliminar')
+				})
+		else:
 			return JsonResponse({
-				"success": True
+					"success": False,
+					"message": ('permiso','No posee los permisos necesarios para realizar para realizar esta operación')
 			})
-		except Error as e:
-			return JsonResponse({
-				"success": False,
-				"message": e.value()
-			})
+	
+	def handle_no_permission(self):
+		messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+		return redirect(self.redirect_url)
 
 ######################################################################
-class GenericListadoView(ExportMixin, SingleTableMixin, FilterView):
-	pass
+#class GenericListadoView(ExportMixin, SingleTableMixin, LoginRequiredMixin,PermissionRequiredMixin,FilterView):
+class GenericListadoView(ExportMixin, SingleTableMixin,FilterView):
+	def handle_no_permission(self):
+		messages.error(self.request, 'No posee los necesarios para realizar permisos para realizar esta operación')
+		return redirect(self.redirect_url)
